pyproject.toml
[project]
name = "un-docs-downloader"
version = "0.1.0"
description = "Download files from the United Nations Official Document System"
requires-python = ">=3.9"
dependencies = [
    "requests>=2.28.0",
    "pymupdf>=1.23.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
]

[project.scripts]
un-docs = "un_docs_downloader.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
]
web = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "python-multipart>=0.0.6",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
pythonpath = ["src"]
